Page 계층 구조로 구현 index->app->page-> (다른 모든 페이지) + Header 부분 page로 빼서 중복 구현 안 해도 되게 만듦
