+++ react-project/src/components/Home.js
@@ -1,0 +1,24 @@
+import Carousel from './Carousel';
+import React from 'react';
+import "../css/Pages.css";
+
+export const Home = () => {
+  return (
+        <>
+          <div className="main_box">
+          <Carousel />
+          </div>
+          </>
+      );
+    };
+
+
+
+
+
+
+
+
+
+
+
