+++ react-project/src/components/Introduce.js
@@ -1,15 +1,11 @@
 import ProductList from "./ProductList";
-
 import React from 'react';
-
-import Header from './Header';
 import "../css/Introduce.css";
 
 
 
 export const Introduce = () => (
     <>
-    <Header />
     <Form />
          <div className="container">
         <ProductList/>
