+++ react-project/src/components/Result.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import Header from './Header';
 import EstimateTable from "./EstimateTable";
 
 
 export const Result = () => (
     <>
-            <Header />
     <div>
         <h1>견적</h1>
         <p>page for 견적</p>
